add asset policy for encrypted notes
The new warning is as follows

WARN: This project uses the default security policy for all assets. While it is set up to work with many applications, it is recommended to further harden the policy to increase security against attacks like XSS.
WARN: To get started, have a look at 'dfx info security-policy'. It shows the default security policy along with suggestions on how to improve it.
WARN: To disable the policy warning, define disable_security_policy_warning: true in .ic-assets.json5.