chore: adds release environment to npm publish workflow (#248)
NPM package publishing will be migrated to use Trusted Publishing (via
OIDC).

To ensure security and restrict access, we will configure the OIDC token
request to only be available to jobs running in the dedicated 'release'
environment. This environment will be a required condition for obtaining
the necessary NPM token.